specify float precision

+++ rave/dataset.py
@@ -61,7 +61,7 @@ class AudioDataset(data.Dataset):
         assert buffer.precision == AudioExample.Precision.INT16
 
         audio = np.frombuffer(buffer.data, dtype=np.int16)
-        audio = audio.astype(np.float) / (2**15 - 1)
+        audio = audio.astype(np.float32) / (2**15 - 1)
 
         if self._transforms is not None:
             audio = self._transforms(audio)
